Fix to work with 10.x.x stable nodejs builds
It looks like this library was developed against the 11.x node.js
branch and has buffer allocaton/handling that isn't in the 10.x
stable series.  This patch allows it to work just with the 10.x
branch.  I will revisit this at some point and make it compile
properly on both.